Switch from float -> double throughout.

+++ src/alchemy.cpp
@@ -31,7 +31,7 @@ bool ReadAlchemy(istream &ifs,OBMol &mol, const char *title)
   ttab.SetFromType("ALC");
 
   string str;
-  float x,y,z;
+  double x,y,z;
   OBAtom *atom;
   vector<string> vs;
