+++ src/chiral.h
@@ -22,11 +22,11 @@ GNU General Public License for more details.
 
 namespace OpenBabel {
 
-void GraphPotentials(OBMol &mol, std::vector<float> &pot);
-void construct_g_matrix(OBMol &mol, std::vector<std::vector<float> > &m);
-void construct_c_matrix(OBMol &mol, std::vector<std::vector<float > > &m);
-float CalcSignedVolume(OBMol &mol,OBAtom*);
-float signed_volume(const vector3 &a, const vector3 &b, const vector3 &c, const vector3 &d);
+void GraphPotentials(OBMol &mol, std::vector<double> &pot);
+void construct_g_matrix(OBMol &mol, std::vector<std::vector<double> > &m);
+void construct_c_matrix(OBMol &mol, std::vector<std::vector<double > > &m);
+double CalcSignedVolume(OBMol &mol,OBAtom*);
+double signed_volume(const vector3 &a, const vector3 &b, const vector3 &c, const vector3 &d);
 void GetChirality(OBMol &mol, std::vector<int> &chirality);
 
 }
